refactored out tooltip content into reusable mixin.
